+++ frontend/src/assets/messageBubble.tsx
@@ -1,6 +1,6 @@
 import SvgIcon, { SvgIconProps } from '@mui/material/SvgIcon';
 
-const ChevronUpIcon = (props: SvgIconProps) => {
+const MessageBubbleIcon = (props: SvgIconProps) => {
   return (
     <SvgIcon>
       <svg
@@ -12,10 +12,10 @@ const ChevronUpIcon = (props: SvgIconProps) => {
         strokeLinecap="round"
         strokeLinejoin="round"
       >
-        <path d="m18 15-6-6-6 6" />
+        <path d="M7.9 20A9 9 0 1 0 4 16.1L2 22Z" />{' '}
       </svg>
     </SvgIcon>
   );
 };
 
-export default ChevronUpIcon;
+export default MessageBubbleIcon;